+++ packages/orq-rc/src/models/operations/retrievedataset.ts
@@ -126,7 +126,7 @@ export const RetrieveDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-02-25T16:39:33.548Z",
+    "2026-02-25T17:35:40.562Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {