+++ packages/orq-rc/RELEASES.md
@@ -7289,3 +7289,13 @@ Based on:
 - [typescript v4.5.0-rc.8] packages/orq-rc
 ### Releases
 - [NPM v4.5.0-rc.8] https://www.npmjs.com/package/@orq-ai/node/v/4.5.0-rc.8 - packages/orq-rc
+
+## 2026-02-25 17:38:49
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.729.0 (2.841.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.5.0-rc.9] packages/orq-rc
+### Releases
+- [NPM v4.5.0-rc.9] https://www.npmjs.com/package/@orq-ai/node/v/4.5.0-rc.9 - packages/orq-rc